+++ changelog.md
@@ -3,6 +3,8 @@
 ## Version 1.1
 
 - Added source file comments to the Java bindings
+- Added git source file information to the generated libraries (for debugging, vc++ only)
+- Generate debug symbol files on all platforms
 
 ## Version 1.0